♻️ Refactor result page
