+++ src/lib/routes.js
@@ -1,8 +1,10 @@
 import { renderIndex } from '../pages'
+import { renderResult } from '../pages/result'
 import { renderAddMenu } from '../pages/add-new'
 
 const routes = {
   '/': renderIndex,
+  '/result': renderResult,
   '/add-menu': renderAddMenu,
 }
